Update welcome.md
On line #37 the word "might" should be changed to "will" to make sense.  The word "occasionally" states that it is an infrequent action, and that action "will" occur occasionally.  Alternatively the sentence could also be changed to "At all times, messages might be delivered in an order different..." but I think the sentence with "Occasionally" is better.  

Thank you for considering my modification.