Update sqs-getting-started.md
I think you should make "queues" singular on line #19, especially since the sentences starts with the phrase "The first".  If the sentence did not have the phrase "The first" and started with "The most" I think either singular or plural would work but the phrase "The first" should be followed with a description of a singular task. 

Thank you for taking the time to review my proposed change.