Update confirm-queue-is-empty.md
`ApproximateNumberOfMessagesVisible` is not a valid property of  `GetQueueAttributes`. I had followed this documentation whilst implementing some logic to poll SQS for final messages and kept getting errors. I had to check the SDK documentation for the language I was using (Go) and realised this was an error. This may require changes in a couple other places in the docs too.